implemented one hot resizing, split classes into seperate files
